header cleanup in aisle binds/python
