+++ binds/python/CMakeLists.txt
@@ -1,4 +1,12 @@
-pybind11_add_module(morphio SYSTEM morphio.cpp)
+pybind11_add_module(morphio
+    SYSTEM
+    morphio.cpp
+    bind_immutable.cpp
+    bindings_utils.cpp
+    bind_misc.cpp
+    bind_mutable.cpp
+    bind_vasculature.cpp
+    )
 
 target_include_directories(morphio
   PRIVATE
