+++ binds/python/bind_immutable.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <pybind11/pybind11.h>
+
+void bind_immutable_module(pybind11::module& m);
