+++ binds/python/bind_misc.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <pybind11/pybind11.h>
+
+void bind_misc(pybind11::module&);
