+++ binds/python/bind_mutable.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <pybind11/pybind11.h>
+
+void bind_mutable_module(pybind11::module& m);
