+++ binds/python/bind_vasculature.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <pybind11/pybind11.h>
+
+void bind_vasculature(pybind11::module&);
