+++ binds/python/morphio.cpp
@@ -1,31 +1,11 @@
 #include <pybind11/pybind11.h>
 
-#include <pybind11/iostream.h>
-#include <pybind11/numpy.h>
-#include <pybind11/operators.h>
-#include <pybind11/stl.h>
-#include <pybind11/stl_bind.h>
-
-#include <morphio/enums.h>
-#include <morphio/types.h>
-
-#include <morphio/errorMessages.h>
-#include <morphio/mito_section.h>
-#include <morphio/mitochondria.h>
-#include <morphio/morphology.h>
-#include <morphio/section.h>
-#include <morphio/soma.h>
-
-#include "bindings_utils.cpp"
-
-#include "bind_immutable.cpp"
-#include "bind_misc.cpp"
-#include "bind_mutable.cpp"
-#include "bind_vasculature.cpp"
+#include "bind_immutable.h"
+#include "bind_misc.h"
+#include "bind_mutable.h"
+#include "bind_vasculature.h"
 
 namespace py = pybind11;
-using namespace py::literals;
-
 
 PYBIND11_MODULE(morphio, m) {
     bind_misc(m);
